packages/advisor_chat/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "wealth_management_portal.advisor_chat" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/packages/advisor_chat/html'
 --cov-report xml:'../../coverage/packages/advisor_chat/coverage.xml'
 --html='../../reports/packages/advisor_chat/unittests/html/index.html'
 --junitxml='../../reports/packages/advisor_chat/unittests/junit.xml'
"""

[tool.hatch.build.targets.wheel]
packages = [ "wealth_management_portal.advisor_chat" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[project]
name = "wealth_management_portal.advisor_chat"
version = "1.0.0"
description = "Advisor Chat: Multi-agent chat system for wealth management advisors"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "wealth_management_portal.common_auth",
  "fastapi==0.128.5",
  "uvicorn==0.40.0",
  "boto3>=1.42.44",
  "strands-agents[a2a,bidi]>=1.30.0",
  "bedrock-agentcore>=0.1.7",
  "yfinance>=1.2.0",
  "python-dotenv>=1.1.0",
  "httpx>=0.28.0",
  "PyJWT[crypto]>=2.8.0",
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

[tool.uv.sources]
"wealth_management_portal.common_auth" = { path = "../common_auth" }

[dependency-groups]
dev = [
    "fastapi[standard]==0.128.5",
    "ruff>=0.11.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.0.0",
    "redshift-connector>=2.1.12",
]
